requirements.txt
pyshorteners
pyperclip
pillow